requirements.txt
anyio==3.7.0
APScheduler==3.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
idna==3.4
oauthlib==3.2.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyevsim==1.1.6
pygsheets==2.0.6
pyparsing==3.0.9
python-telegram-bot==20.3
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tqdm==4.65.0
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
